+++ src/main/java/web/mvc/service/ReplyServiceImpl.java
@@ -3,6 +3,8 @@ package web.mvc.service;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import web.mvc.domain.Reply;
+import web.mvc.exception.BasicException;
+import web.mvc.exception.ErrorCode;
 import web.mvc.repository.ReplyRepository;
 
 @Service
@@ -18,6 +20,10 @@ public class ReplyServiceImpl implements ReplyService {
 
     @Override
     public void delete(Long id) {
+        if(!replyRepository.existsById(id)) {
+            throw new BasicException(ErrorCode.NOTFOUND_ID);
+        }
+
         replyRepository.deleteById(id);
     }
 }
